polymarket-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
httpx==0.25.2
python-dotenv==1.0.0
py-clob-client==0.17.0
pydantic==2.5.2
aiohttp==3.9.1
